Fix python in recipe
pymoab doesn't work under python>3.6 as the recipe is written.
I don't see anything in the moab repo stating that newer python
isn't supported.  It is getting built ostensibly with newer
python but pymoab is ending up in the python3.6 folder.

Add xylar as a maintainer

Add more tests that would have uncovered the problems with python